+++ src/vyper/mod.rs
@@ -43,11 +43,12 @@ impl Compiler {
     pub const DEFAULT_EXECUTABLE_NAME: &'static str = "vyper";
 
     /// The supported versions of `vyper`.
-    pub const SUPPORTED_VERSIONS: [semver::Version; 4] = [
+    pub const SUPPORTED_VERSIONS: [semver::Version; 5] = [
         semver::Version::new(0, 3, 3),
         semver::Version::new(0, 3, 9),
         semver::Version::new(0, 3, 10),
         semver::Version::new(0, 4, 0),
+        semver::Version::new(0, 4, 1),
     ];
 
     /// The first version where we cannot use the optimizer.