+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for `vyper==0.4.1`
+- Update tests to include `vyper==0.4.1` and others
+- Add Vyper binary for `0.4.1`
+
 ### Changed
 
 - Updated to Rust v1.85.1