+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Set the stack size limit for all threads to 64 MB
 - Updated to Rust v1.85.1
 
 ## [1.5.9] - 2025-01-09